chore(dynamodb): explicitly render `false` pointInTimeRecovery
Today, if the property `pointInTimeRecoveryEnabled` is either `undefined`, or `false`,
we don't render the `pointInTimeRecoverySpecification` property in the Table L1.
A customer wants to write an Aspect that checks whether this property has been set,
and they cannot differentiate between the "not provided" and "set to `false`" cases.

Change the logic to render `pointInTimeRecoverySpecification` even for `false`
pointInTimeRecoveryEnabled values.

Fixes #19748