fix(rds): MySQL 8.0 uses wrong Parameter for S3 export
In the newest version of the MySQL 8.0 Aurora engine,
the S3 import and export Roles need to be combined,
which is the first time this is needed for a Cluster
(previously, only Instances combined those two Roles).
Introduce this concept as a first-class property of Cluster engine,
`false` by default, and make it `true` in MySQL version 8.0.

Fixes #19735