fix(cli): stack monitor prints over errors
In #19742, we limited the amount of text that the status printer
would produce, and tried to get rid of the empty lines at the end
of the rewritable block as well.

Unfortunately, we scrolled up too far, overwriting the error messages
that may occur during resource provisioning. Since these were the only
place where error messages were displayed, it is now not possible to see
why a stack deployment fails.

This PR fixes that.