chore(cli): include resource errors in exception
The resource errors that were encountered during the deployment are very
important, and the main errors the users should be paying attention to.

They weren't necessarily being given the front and center stage they
needed to though. Usually, users would see the following closing error
message:

```
The stack named <Stack> failed to deploy: UPDATE_ROLLBACK_COMPLETE
```

Leaving out the actual error messages.

Record them as part of the monitor, and add them to the final exception.

(Unfortunately, no automated tests as this part of the code isn't really
set up to be testable and I don't think investing the effort to make it
is worth it right now).