feat(cli): allow diffing against a processed template
In some cases, using an unprocessed template with `CfnInclude` fails,
most commonly because the unprocessed template can contain cycles between the resources.
When this happens, customers need to include the processed template instead,
but when that happens, the `diff` command returns the wrong results,
because it compares against the original, and thus unprocessed, template.

Add a new command-line switch to the `diff` command that allows comparing against the processed template,
which gives the correct results in this case.