fix(cfn-include): allow CFN Functions in Tags
Our `TagManger` does not allow expressing Tags that included top-level CloudFormation functions,
like `Fn::If`. While that's not a big issue when writing CDK code directly,
it makes it impossible to include some CloudFormation templates that use that pattern.

Introduce the concept of "dynamic tags" to the `TagManager` that allows preserving these,
and return them alongside the "regular" Tags when rendering.

Fixes #16889