chore: tests don't run on Linux system (#19989)
The first argument to `mkdtemp` is a *prefix*, not a *parent directory*.

In other words, on a regular Linux system the following code:

```
mkdtemp(os.tmpdir())
```

Tries to make a *sibling* of the temp directory instead of a child (for
example, `/tmpABC123` instead of `/tmp/ABC123`), which then fails with
"access denied".

This doesn't fail on Mac because the `tmp` directory is in another
directory with write permissions so we can create a sibling directory.
And in CodeBuild we run as root so we have permissions to create a
sibling dir in the root.

Fix it for other systems.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*