fix: deploy monitor count is off if there are > 100 changes
The deployment monitor failed to account for the fact the
`DescribeChangeSet` operation returns a paginated output, and only
represents up to 100 changes per pages.

Since only the first page was considered, the progress bar would always
start with a value of 101 (or less), instead of the correct count.

This aggregates the `Changes` arrays from all pages of the change set
description so the count is correct.