feat(cli): parse future flag context arguments as boolean
All future flag context arguments are either `true` or `false` (enabled
or disabled). When passing future flags as command line arguments, i.e.

```
cdk deploy --context @aws-cdk/core:newStyleStackSynthesis=false
cdk deploy --context @aws-cdk/core:newStyleStackSynthesis=0
```

The values are interpreted as strings so the above two values get
interpreted as 'false' and '0' which are truthy (and not what we want).

This PR adds a function to parse context arguments and covert any future
flags to boolean values. 'false' and '0' will be evaluated as `false`.

This functionality could be beneficial to more than just future flags,
but with future flags we know that the value must be a boolean.