fix: deploy monitor count is off if there are > 100 changes (#20067)
The deployment monitor failed to account for the fact the
`DescribeChangeSet` operation returns a paginated output, and only
represents up to 100 changes per pages.

Since only the first page was considered, the progress bar would always
start with a value of 101 (or less), instead of the correct count.

This aggregates the `Changes` arrays from all pages of the change set
description so the count is correct.

Fixes #11805

----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*