fix(integ-runner): only diff registered stacks
Currently the runner will perform a diff of all stacks in the snapshot,
not just the stacks that have been "registered" as part of the test
case. This PR changes the logic to only perform a diff on stacks that
are part of the test case. This will allow us to use separate stacks for
things like assertions that will not be part of the diff.

As part of this change it was necessary to make some additional changes
to the way the runner runs the tests, the main reason being that it is
necessary to hold two different versions of the test, the "expected"
version and the "actual" version.

This PR updates the runner to read the "expected" version of the test
from the `integ.json` manifest in the current snapshot and the "actual"
version of the test from the changed integ test. Because the legacy
tests (those that don't use the `IntegTest` construct) do not create an
`integ.json` file, this PR updates the runner to create an `integ.json`
file as part of the snapshot creation. This way on the next change the
runner is able to read the `integ.json` file from the snapshot to get
the "expected" view of the test.