feat(cloudwatch): Add CustomWidget (#19327)
This PR introcudes the CustomWidget for Cloudwatch Dashboards

fixes: #17579

based on the work of @dannyber

If you wonder, why only a string typed option was used instead of an lambda.IFunction, the reason is, that this would introduce a circular dependency between lambda -> cloudwatch -> lambda and leads to build errors.

Creating a new package for this tiny feature to work around the circular dependency would be too much of a good thing. Apart from that it could be unintuitive to find the custom widget in a package 'aws-cloudwatch-pattarns' if all other widgets come from 'aws-cloudwatch'.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*