fix(ecs-patterns): feature flag missing for breaking change passing container port for target group port (#20284)
PR #18157 results in a new TargetGroup being created from NetworkLoadBalancedEc2Service, NetworkLoadBalancedFargateService, NetworkMultipleTargetGroupsEc2Service,
and NetworkMultipleTargerGroupsFargateService even when no change is made because we are now passing through the containerPort props to TargetGroups's Port.

For existing services, this is a breaking change so a feature flag is needed. This PR adds that feature flag.

Closes #19411.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*