chore(integ-runner): `--inspect-failures` option
When the integ runner fails to validate a snapshot, add an option
to retain the tempdir with the "new actual" snapshot so that a human
can go look around in it.

Also make the `--verbose` flag print out a command that will reproduce
the running of the CDK app directly from the command-line, so that
it can be debugged more easily.