chore: add reprioritized issues to ownership project board (#20386)
The current workflow for re-prioritization is to automatically comment on the issue that a CDK maintainer will provide an update. But we do not track that anywhere. I added a feature to the github action that will add re-prioritized issues to a repository project. This update aligns the workflow with the new feature.

See the relevant [readme](https://github.com/kaizencc/issue-reprioritization-manager#add-reprioritized-issues-to-a-github-project) for the `project-column-url` property for more. `repository-projects: write` allows github actions permission to add new cards to the project.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*