chore: fix command to run integ tests
The instructions in `CONTRIBUTING.md`, telling users to run `test.sh`,
started running the unit tests instead of the integ tests, since `npx`
now `cd`s to the `package.json` directory (instead of running in the
current directory).

Fix by using `npx` to look up the location of `jest`, but running
it in the current shell.