feat(core): allow disabling of LogicalID Metadata in case of large manifest (#20433)
Users have encountered an error resulting from the manifest being too large to stringify. This allows users to prevent this metadata from ever being added to the manifest. Fixes the issue that caused the previous iteration of this PR to fail in the v2 pipeline.

Fixes https://github.com/aws/aws-cdk/issues/20211.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*