chore(integ-runner): print long-running snaphots
Print snapshot validations that are running for a long time (1 minute).
This helps diagnose what is stuck, if anything is. On my machine, it
was tests using Docker because there was some issue with it, and this
lost me a day.

Also change the test reporting formatting slightly, and catch errors
during actual test running.