chore: update dependabot and yarn-upgrade config (#19942)
Closes https://github.com/cdklabs/cdk-ops/issues/1942

The idea is to copy the configuration we already have for the default branch and target/checkout `v1-main` on June 1. 
It is a bit ugly but -- we are not likely to do this again to any other branch, and I can't think of a better way. I have tested the changes on my own forked repository and confirmed that we do indeed target/checkout a different branch under the new configuration.

Before this PR is merged, we should double-check and make sure that yarn-update has not diverged. Targeting `v2-main` as that will be the default branch when this PR gets merged (I assume the conflicts will be resolved when `v2-main` gets forward merged).

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*